requirements.txt
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==4.4.0 ; python_version >= "3.12" and python_version < "4.0"
cachetools==5.3.3 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.6.2 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.12" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and sys_platform == "win32" or python_version >= "3.12" and python_version < "4.0" and platform_system == "Windows"
dnspython==2.6.1 ; python_version >= "3.12" and python_version < "4.0"
email-validator==2.2.0 ; python_version >= "3.12" and python_version < "4.0"
fastapi-cli==0.0.4 ; python_version >= "3.12" and python_version < "4.0"
fastapi==0.111.0 ; python_version >= "3.12" and python_version < "4.0"
google-ai-generativelanguage==0.6.6 ; python_version >= "3.12" and python_version < "4.0"
google-api-core==2.19.1 ; python_version >= "3.12" and python_version < "4.0"
google-api-core[grpc]==2.19.1 ; python_version >= "3.12" and python_version < "4.0"
google-api-python-client==2.135.0 ; python_version >= "3.12" and python_version < "4.0"
google-auth-httplib2==0.2.0 ; python_version >= "3.12" and python_version < "4.0"
google-auth==2.30.0 ; python_version >= "3.12" and python_version < "4.0"
google-generativeai==0.7.1 ; python_version >= "3.12" and python_version < "4.0"
googleapis-common-protos==1.63.2 ; python_version >= "3.12" and python_version < "4.0"
grpcio-status==1.62.2 ; python_version >= "3.12" and python_version < "4.0"
grpcio==1.64.1 ; python_version >= "3.12" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.12" and python_version < "4.0"
httpcore==1.0.5 ; python_version >= "3.12" and python_version < "4.0"
httplib2==0.22.0 ; python_version >= "3.12" and python_version < "4.0"
httptools==0.6.1 ; python_version >= "3.12" and python_version < "4.0"
httpx==0.27.0 ; python_version >= "3.12" and python_version < "4.0"
idna==3.7 ; python_version >= "3.12" and python_version < "4.0"
jinja2==3.1.4 ; python_version >= "3.12" and python_version < "4.0"
markdown-it-py==3.0.0 ; python_version >= "3.12" and python_version < "4.0"
markupsafe==2.1.5 ; python_version >= "3.12" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.12" and python_version < "4.0"
orjson==3.10.5 ; python_version >= "3.12" and python_version < "4.0"
proto-plus==1.24.0 ; python_version >= "3.12" and python_version < "4.0"
protobuf==4.25.3 ; python_version >= "3.12" and python_version < "4.0"
pyasn1-modules==0.4.0 ; python_version >= "3.12" and python_version < "4.0"
pyasn1==0.6.0 ; python_version >= "3.12" and python_version < "4.0"
pydantic-core==2.18.4 ; python_version >= "3.12" and python_version < "4.0"
pydantic-settings==2.3.4 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.7.4 ; python_version >= "3.12" and python_version < "4.0"
pygments==2.18.0 ; python_version >= "3.12" and python_version < "4.0"
pyparsing==3.1.2 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
python-multipart==0.0.9 ; python_version >= "3.12" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.12" and python_version < "4.0"
rich==13.7.1 ; python_version >= "3.12" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.12" and python_version < "4"
shellingham==1.5.4 ; python_version >= "3.12" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
starlette==0.37.2 ; python_version >= "3.12" and python_version < "4.0"
tqdm==4.66.4 ; python_version >= "3.12" and python_version < "4.0"
typer==0.12.3 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.12" and python_version < "4.0"
ujson==5.10.0 ; python_version >= "3.12" and python_version < "4.0"
uritemplate==4.1.1 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.2.2 ; python_version >= "3.12" and python_version < "4.0"
uvicorn[standard]==0.30.1 ; python_version >= "3.12" and python_version < "4.0"
uvloop==0.19.0 ; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_version >= "3.12" and python_version < "4.0"
watchfiles==0.22.0 ; python_version >= "3.12" and python_version < "4.0"
websockets==12.0 ; python_version >= "3.12" and python_version < "4.0"
